reserve functionality updated(many)
edited reserve with temporary user functionality, if current user is guest, goes to login page when he tries to reserve.

only allow user to reserve only one copy of a certain book x. This is done by adding a reservers array field inside booksdemo documents, which registers the username currently login>>

in the actual output, the username inside reservers array should be replaced with userid to have a unique identity...

the loginpage component has function called validate, which validates the username and password input, instead of currUser being set to username, set to userid instead...

userpage(tried but not finished)...to make it such that the login link inside navbar would change to whatever username is currently login, as well as its actual to component